refactor: isolate legacy yaml packages Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
