+++ internal/legacyyaml/migrate/migrate.go
@@ -12,7 +12,7 @@ import (
 	"github.com/grafana/oats/casefile"
 	"github.com/grafana/oats/discovery"
 	"github.com/grafana/oats/internal/legacyyaml"
-	"github.com/grafana/oats/model"
+	"github.com/grafana/oats/internal/legacyyaml/model"
 	goyaml "go.yaml.in/yaml/v3"
 )
