+++ internal/legacyyaml/migrate/migrate_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/grafana/oats/casefile"
 	"github.com/grafana/oats/discovery"
-	"github.com/grafana/oats/model"
+	"github.com/grafana/oats/internal/legacyyaml/model"
 	"github.com/grafana/oats/testhelpers/kubernetes"
 )
 
@@ -104,7 +104,7 @@ func TestConvertDefinition_SingleMatrixComposeRequiresFile(t *testing.T) {
 }
 
 func TestConvertFile_RendersYAML(t *testing.T) {
-	sample := filepath.Join("..", "internal", "legacyyaml", "testdata", "valid-tests", "oats.yaml")
+	sample := filepath.Join("..", "testdata", "valid-tests", "oats.yaml")
 	out, warnings, err := ConvertFile(sample)
 	if err != nil {
 		fatalf(t, "ConvertFile: %v", err)
@@ -194,7 +194,7 @@ func TestMigrateHelpers(t *testing.T) {
 }
 
 func TestConvertFile_MatrixSampleIsParseable(t *testing.T) {
-	sample := filepath.Join("..", "internal", "legacyyaml", "testdata", "valid-tests", "matrix-test.oats.yaml")
+	sample := filepath.Join("..", "testdata", "valid-tests", "matrix-test.oats.yaml")
 	out, warnings, err := ConvertFile(sample)
 	if err != nil {
 		fatalf(t, "ConvertFile matrix sample: %v", err)
