+++ internal/legacyyaml/prometheus/responses/response_helpers.go
