+++ internal/legacyyaml/validate_test.go
@@ -3,7 +3,7 @@ package legacyyaml
 import (
 	"testing"
 
-	"github.com/grafana/oats/model"
+	"github.com/grafana/oats/internal/legacyyaml/model"
 	"github.com/onsi/gomega"
 	"github.com/stretchr/testify/require"
 )
